Avoid adding two prices in hledger
beancount2ledger converts the cost to a price, which means that we
can end up with two price info in a posting, which is invalid syntax.

For now, stop adding a second price; however, according to the
hledger FAQ at https://hledger.org/faq.html hledger supports the
lot syntax now (even though it doesn't use the information) so
it might be best to use that.

Fixes #2