Add rounding postings only when required
Ledger and beancount handle rounding and precision differently.
In ledger (and hledger), the entry with the most precision (in
the whole file) is taken for a currency.  This means that an
unrelated transaction can influence whether we need a rounding
entry for a transaction.

To solve this, we build up a dcontext based on all entries and
use that to format amounts -- if the formatted amount of a
rounding posting is zero, we don't have to add it (and in fact
adding it might break ledger).

However, for costs and prices we always have to retain the full
precision.

Fixes #9