Avoid creating two or more postings with null amount
If a posting without an amount is given and several amounts would
be added when balancing, beancount will create several postings.
But we ignore the amount on those postings (since they were added
by beancount and not the user), which means we may end up with
two or more postings with no amount, which is not valid.

Therefore, only take *one* posting by looking at the line number.

This fixes a regression which was introduced in commit 27dcd7eec0
("Show amounts only when they are specified in the source file").

Fixes #23