Add price to cost when needed by ledger
Ledger doesn't use the cost to determine an exchange (see
https://github.com/ledger/ledger/issues/630), which means
we have to add the price explicitly if ledger can't do
an implicit conversion.

This improves the logic that was added in commit e915d2b7
("Support entries mixed with held-at-cost and price-conversion
postings") to deal with one test case.

Fixes #22